ugly commit - working parkour geometry and lights
Split VectorHelper up.
Add GTest and get some tests working.
Added SkyboxShape and SimpleRenderShape.
Change DeltaTime to a Time class.
Add Menu UI.

+++ src/Engine/Drawing/BindingManager.cpp
@@ -12,7 +12,7 @@ namespace Drawing
 	{
 		GLint maxUniforms = 0;
 		glGetIntegerv(GL_MAX_UNIFORM_BUFFER_BINDINGS, &maxUniforms);
-		if (_next >= maxUniforms)
+		if (_next >= (GLuint)maxUniforms)
 		{
 			DERROR("Exceeded maximum number of uniform buffer bindings!");
 			throw std::runtime_error("Ran out of OpenGL Uniform Buffer Bindings!");
